requirements.txt

python-dotenv
requests
pydantic
beautifulsoup4
openai
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.33.0
mlflow>=2.14.0
sentencepiece
tiktoken
torch>=2.6.0
torchvision
torchaudio
streamlit